+++ dev/sparktestsupport/modules.py
@@ -750,7 +750,11 @@ pyspark_pandas = Module(
         "pyspark.pandas.tests.resample.test_series",
         "pyspark.pandas.tests.resample.test_timezone",
         "pyspark.pandas.tests.test_reshape",
-        "pyspark.pandas.tests.test_rolling",
+        "pyspark.pandas.tests.window.test_rolling",
+        "pyspark.pandas.tests.window.test_rolling_error",
+        "pyspark.pandas.tests.window.test_groupby_rolling",
+        "pyspark.pandas.tests.window.test_groupby_rolling_adv",
+        "pyspark.pandas.tests.window.test_groupby_rolling_count",
         "pyspark.pandas.tests.test_scalars",
         "pyspark.pandas.tests.test_series_conversion",
         "pyspark.pandas.tests.test_series_datetime",
@@ -1118,7 +1122,11 @@ pyspark_pandas_connect_part2 = Module(
         "pyspark.pandas.tests.connect.window.test_parity_ewm_error",
         "pyspark.pandas.tests.connect.window.test_parity_ewm_mean",
         "pyspark.pandas.tests.connect.window.test_parity_groupby_ewm_mean",
-        "pyspark.pandas.tests.connect.test_parity_rolling",
+        "pyspark.pandas.tests.connect.window.test_parity_rolling",
+        "pyspark.pandas.tests.connect.window.test_parity_rolling_error",
+        "pyspark.pandas.tests.connect.window.test_parity_groupby_rolling",
+        "pyspark.pandas.tests.connect.window.test_parity_groupby_rolling_adv",
+        "pyspark.pandas.tests.connect.window.test_parity_groupby_rolling_count",
         "pyspark.pandas.tests.connect.test_parity_expanding",
         "pyspark.pandas.tests.connect.test_parity_ops_on_diff_frames_groupby_rolling",
         "pyspark.pandas.tests.connect.computation.test_parity_missing_data",
