+++ python/pyspark/pandas/tests/connect/window/test_parity_rolling.py
@@ -16,19 +16,22 @@
 #
 import unittest
 
-from pyspark.pandas.tests.test_rolling import RollingTestsMixin
+from pyspark.pandas.tests.window.test_rolling import RollingMixin
 from pyspark.testing.connectutils import ReusedConnectTestCase
 from pyspark.testing.pandasutils import PandasOnSparkTestUtils, TestUtils
 
 
 class RollingParityTests(
-    RollingTestsMixin, PandasOnSparkTestUtils, TestUtils, ReusedConnectTestCase
+    RollingMixin,
+    PandasOnSparkTestUtils,
+    TestUtils,
+    ReusedConnectTestCase,
 ):
     pass
 
 
 if __name__ == "__main__":
-    from pyspark.pandas.tests.connect.test_parity_rolling import *  # noqa: F401
+    from pyspark.pandas.tests.connect.window.test_parity_rolling import *  # noqa: F401
 
     try:
         import xmlrunner  # type: ignore[import]
